+++ lib/layrz_theme.dart
@@ -28,3 +28,4 @@ export 'src/utilities/utilities.dart';
 export 'src/map/map.dart';
 export 'src/colorblindness/colorblindness.dart';
 export 'src/alerts/alerts.dart';
+export 'src/chips/chips.dart';